+++ lib/plugins/worker-compiler.js
@@ -8,36 +8,6 @@ var fs = require('fs');
 var iterateWorkers = require('../helpers/iterate-workers');
 var debug = require('debug')('skyrocket');
 var chalk = require('chalk');
-var EOL = require("os").EOL;
-
-function build(file, inputs) {
-  inputs.forEach(function (input) {
-    fs.appendFileSync(file, input + EOL);
-  });
-  return true;
-}
-
-function createWorkerShell(path, name) {
-  var lines = [
-    "import WorkerApp from './worker';",
-    "import Contract from './contract';",
-    "import WorkerShell from 'skyrocket/-private/-worker-shell';",
-    "",
-    "export default WorkerShell.extend({",
-    "  worker: WorkerApp",
-    "  contract: Contract",
-    "  name:" + name,
-    "});"
-  ];
-  return build(path, lines);
-}
-
-function createWorkerBoot(path) {
-  var lines = [
-    "var WorkerInstance = require('" + path + "')['default'].create({});"
-  ];
-  return build(path, lines);
-}
 
 // Create a subclass from Plugin
 WorkerCompiler.prototype = Object.create(Plugin.prototype);
