+++ lib/plugins/worker-support/ember-shim.js
@@ -1,0 +1,1 @@
+export default Ember;
