+++ lib/plugins/worker-support/prefix/shims.js
@@ -1,0 +1,12 @@
+
+define('ember', [], function() {
+  'use strict';
+
+  return { 'default': Ember };
+});
+
+define('rsvp', [], function() {
+  'use strict';
+
+  return { 'default': Ember.RSVP, 'Promise': Ember.RSVP.Promise };
+});
